Move OrganisationLookup out of ContentAPI module
The module name is no longer appropriate as the lookup deals
with more than just the Content API.
