+++ app/models/problem_report.rb
@@ -1,5 +1,4 @@
 require 'csv'
-require 'content_api/organisation_lookup'
 
 class ProblemReport < AnonymousContact
   validates :what_doing, length: { maximum: 2 ** 16 }
